refactor(ProductService Directory)
