+++ src/main/java/kitchenpos/products/tobe/domain/ProductPrice.java
@@ -8,7 +8,8 @@ class ProductPrice {
     @Column(name = "price", nullable = false)
     private Long price;
 
-    protected ProductPrice() {}
+    protected ProductPrice() {
+    }
 
     public ProductPrice(final Long price) {
         validate(price);
@@ -16,7 +17,7 @@ class ProductPrice {
     }
 
     private void validate(final Long price) {
-        if(price == null || price < 0) {
+        if (price == null || price < 0) {
             throw new IllegalArgumentException("제품 가격은 0원 이상이여야한다.");
         }
     }
