+++ src/main/java/kitchenpos/products/tobe/domain/ProductRepository.java
@@ -2,6 +2,6 @@ package kitchenpos.products.tobe.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Product, Long>{
+public interface ProductRepository extends JpaRepository<Product, Long> {
 
 }
