+++ src/main/java/kitchenpos/products/tobe/domain/ProductService.java
@@ -1,7 +1,5 @@
-package kitchenpos.products.tobe.application;
+package kitchenpos.products.tobe.domain;
 
-import kitchenpos.products.tobe.domain.Product;
-import kitchenpos.products.tobe.domain.ProductRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
